Backend/requirement.txt
django-cors-headers
djangorestframework
django
psycopg2
python-decouple
djangorestframework-simplejwt
